refactor(services): create shared/ subpackage and migrate utility functions (#2392)
Phase 1-2a: Create services/shared/ subpackage with lazy-import pattern
and migrate 5 utility files with backward-compatible re-export bridges.

Changes:
- Created services/shared/ subpackage with __init__.py lazy-import pattern
- Merged path_helpers.py + git_path_client.py → shared/paths.py
- Migrated error_helpers.py → shared/errors.py
- Migrated label_utils.py → shared/labels.py
- Migrated branch_arg.py → shared/branches.py
- Updated handoff_resolution.py to import from shared/paths
- Created re-export bridges in all 5 old files for backward compatibility
- Updated services/__init__.py _SYMBOL_MODULES and TYPE_CHECKING imports
- Updated docs references to use new module paths

Verification:
- Type checking: mypy PASS
- Lint: ruff PASS
- Core tests: 53 tests PASS (path_helpers, git_path_client, label_utils)
- Mock patch compatibility: Some test failures expected per plan Risk #2,
  to be addressed in future phases

All external consumers (from vibe3.services import X) continue working
without changes via _SYMBOL_MODULES mapping and re-export bridges.
