+++ src/vibe3/services/handoff_resolution.py
@@ -23,7 +23,7 @@ import re
 import tomllib
 from pathlib import Path
 
-from vibe3.services.git_path_client import (
+from vibe3.services.shared.paths import (
     GitPathProtocol,
     _get_git_client,
     find_worktree_path_for_branch,
